move get to handler
